+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Features:
 
 - text overlays
 - watermarks and text overlays positioning using gravity
+- Google Cloud Storage support (by @hjr265)